services/gemini-proxy/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai
python-multipart
